Rl with clutter map
the reward function has been changed accrodingly.

if object is not seen that the white_pixel__cout is 0 then the difference the difference between the previous and current total clutter density is calculated. this difference is subtracted from the reward in order to motivate the robot to reduce the total clutter density

when an object is seen and the centroid is not -1,1 then a local clutter density is defined around a radius of 40 from the centroid, then the difference the difference between the previous and current local clutter density is calculated. this difference is subtracted from the reward in order to motivate the robot to reduce the  clutter density around the target object which is the centroid